requirement.txt
fastapi
uvicorn
psycopg2-binary
python-jose
passlib
pydantic-settings
apscheduler
pytz
